some progress

+++ hermes2d/test_examples/FCT/DG_CG/Partikel/test/interface.h
@@ -54,7 +54,7 @@ protected:
 
 			NumericalFlux* num_flux;
 			double gamma;
-			double entry_i;
+			int entry_i;
 
   };
